// apps/api/src/id-business-v2/workspace/dto/id-business-v2-relay-script.dto.ts
export interface SaveIdBusinessV2RelayGoogleOAuthDto {
  clientId?: unknown;
  clientSecret?: unknown;
}

export interface LoginIdBusinessV2RelayCloudBridgeDto {
  email?: unknown;
  password?: unknown;
  tempToken?: unknown;
  totpCode?: unknown;
}

export interface CreateIdBusinessV2RelayJobDto {
  accountConcurrency?: unknown;
  accountLabel?: unknown;
  accountLoadFactor?: unknown;
  accountPriority?: unknown;
  accountRateMultiplier?: unknown;
  allowOverages?: unknown;
  apiKey?: unknown;
  billingAccount?: unknown;
  creditExpiresAt?: unknown;
  deploymentKey?: unknown;
  googleEmail?: unknown;
  location?: unknown;
  mode?: unknown;
  mixedScheduling?: unknown;
  projectDisplayName?: unknown;
  projectId?: unknown;
  proxyId?: unknown;
  referenceAccountId?: unknown;
  selectedModels?: unknown;
  targetGroupId?: unknown;
}

export interface CompleteIdBusinessV2RelaySubscriptionAuthorizationDto {
  callbackUrl?: unknown;
}